Remove the integration test and place it in the unit test file as suggested.
